# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for Auth Service.",
        "title": "Auth Service API3",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:4444",
    "paths": {
...
        "/swap/adoptionSwap": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ushbu endpoint asosida swap so'rovi uchun sababni qabul qilish uchun ishlatiladi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "swap"
                ],
                "summary": "Swap so'rovini qabul qilish",
                "parameters": [
                    {
                        "description": "Reason Payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swaps.Reason"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/swaps.Responce"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/swap/allRequests/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ushbu endpoint asosida barcha swap so'rovlarni olish uchun ishlatiladi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "swap"
                ],
                "summary": "Barcha swap so'rovlarni olish",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Status of swap requests (e.g., pending, accepted, rejected)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Limit of items to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/swaps.AllSwaps"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/swap/rejectRequest": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ushbu endpoint asosida swap so'rovini rad qilish uchun sababni qabul qilish uchun ishlatiladi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "swap"
                ],
                "summary": "Swap so'rovini rad qilish",
                "parameters": [
                    {
                        "description": "Reason Payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swaps.Reason"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/swaps.Responce"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/swap/sendRequest": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ushbu endpoint avval kiritilgan ma'lumotlar bilan swap so'rovini jo'natish uchun ishlatiladi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "swap"
                ],
                "summary": "Swap so'rovini jo'natish",
                "parameters": [
                    {
                        "description": "Swap Request Payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swaps.SwapRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/swaps.SwapResponce"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "swaps.AllSwaps": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "swaps": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/swaps.Swap"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "swaps.Reason": {
            "type": "object",
            "properties": {
                "reason": {
                    "type": "string"
                },
                "swapId": {
                    "type": "string"
                }
            }
        },
        "swaps.Responce": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "offeredItemId": {
                    "type": "string"
                },
                "ownerId": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "requestedItemId": {
                    "type": "string"
                },
                "requesterId": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "swaps.Swap": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "offeredItemId": {
                    "type": "string"
                },
                "ownerId": {
                    "type": "string"
                },
                "requestedItemId": {
                    "type": "string"
                },
                "requesterId": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "swaps.SwapRequest": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "offeredItemId": {
                    "type": "string"
                },
                "ownerId": {
                    "type": "string"
                },
                "requestedItemId": {
                    "type": "string"
                },
                "requesterId": {
                    "type": "string"
                }
            }
        },
        "swaps.SwapResponce": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "offeredItemId": {
                    "type": "string"
                },
                "ownerId": {
                    "type": "string"
                },
                "requestedItemId": {
                    "type": "string"
                },
                "requesterId": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}